Move serializer and exception to notifers folder
