+++ app/jobs/exception_hunter/send_notification_job.rb
@@ -4,7 +4,8 @@ module ExceptionHunter
 
     def perform(serialized_notifier)
       # Use SlackNotifierSerializer as it's the only one for now.
-      deserialized_notifier = ExceptionHunter::Serializers::SlackNotifierSerializer.deserialize(serialized_notifier)
+      serializer = ExceptionHunter::Notifiers::Serializers::SlackNotifierSerializer
+      deserialized_notifier = serializer.deserialize(serialized_notifier)
       deserialized_notifier.notify
     rescue Exception # rubocop:disable Lint/RescueException
       # Suppress all exceptions to avoid loop as this would create a new error in EH.
