+++ lib/exception_hunter.rb
@@ -8,8 +8,8 @@ require 'exception_hunter/error_reaper'
 require 'exception_hunter/tracking'
 require 'exception_hunter/user_attributes_collector'
 require 'exception_hunter/notifiers/slack_notifier'
-require 'exception_hunter/serializers/slack_notifier_serializer'
-require 'exception_hunter/exceptions/misconfigured_notifiers.rb'
+require 'exception_hunter/notifiers/serializers/slack_notifier_serializer'
+require 'exception_hunter/notifiers/exceptions/misconfigured_notifiers.rb'
 
 module ExceptionHunter
   autoload :Devise, 'exception_hunter/devise'
@@ -34,7 +34,7 @@ module ExceptionHunter
     notifiers.each do |notifier|
       next if notifier[:name] == :slack && notifier.dig(:options, :webhook).present?
 
-      raise ExceptionHunter::Exceptions::MisconfiguredNotifiers, notifier
+      raise ExceptionHunter::Notifiers::Exceptions::MisconfiguredNotifiers, notifier
     end
   end
 end
