--- lib/exception_hunter/exceptions/misconfigured_notifiers.rb
@@ -1,9 +1,0 @@
-module ExceptionHunter
-  module Exceptions
-    class MisconfiguredNotifiers < StandardError
-      def initialize(notifier)
-        super("Notifier has incorrect configuration: #{notifier.inspect}")
-      end
-    end
-  end
-end
