+++ lib/exception_hunter/notifiers/exceptions/misconfigured_notifiers.rb
@@ -1,0 +1,11 @@
+module ExceptionHunter
+  module Notifiers
+    module Exceptions
+      class MisconfiguredNotifiers < StandardError
+        def initialize(notifier)
+          super("Notifier has incorrect configuration: #{notifier.inspect}")
+        end
+      end
+    end
+  end
+end
